test: move Redis lifecycle into Vitest globalSetup
