+++ vitest.config.ts
@@ -3,8 +3,8 @@ import { defineConfig } from 'vitest/config'
 export default defineConfig({
   test: {
     setupFiles: ['./tests/setup.ts'],
-    // tests/integration/redis/* need a live Redis server (started by
-    // scripts/run-integration-tests.mjs), so they only run via vitest.integration.config.ts.
+    // tests/integration/redis/* need a live Redis server (globalSetup in
+    // vitest.integration.config.ts), so they only run via that config.
     exclude: ['**/node_modules/**', '**/dist/**', '**/docs-web/**', 'tests/integration/redis/**'],
     coverage: {
       provider: 'v8',
